+++ src/wrappers/uavcan/primitive/scalar/Integer32_1_0.hpp
@@ -1,12 +1,12 @@
 /**
  * This software is distributed under the terms of the MIT License.
  * Copyright (c) 2020 LXRobotics.
- * Author: Alexander Entinger <alexander.entinger@lxrobotics.com>
+ * Author: Bernhard Mayer <bernard@generationmake.de>, Alexander Entinger <alexander.entinger@lxrobotics.com>
  * Contributors: https://github.com/107-systems/107-Arduino-UAVCAN/graphs/contributors.
  */
 
-#ifndef ARDUINO_UAVCAN_TYPES_REG_DRONE_PHYSICS_ELECTRICITY_POWER_0_1_H_
-#define ARDUINO_UAVCAN_TYPES_REG_DRONE_PHYSICS_ELECTRICITY_POWER_0_1_H_
+#ifndef ARDUINO_UAVCAN_TYPES_UAVCAN_PRIMITIVE_SCALAR_INTEGER32_1_0_HPP_
+#define ARDUINO_UAVCAN_TYPES_UAVCAN_PRIMITIVE_SCALAR_INTEGER32_1_0_HPP_
 
 /**************************************************************************************
  * INCLUDE
@@ -14,56 +14,54 @@
 
 #include <libcanard/canard.h>
 
-#include "Power_0_1.h"
+#include <types/uavcan/primitive/scalar/Integer32_1_0.h>
 
 /**************************************************************************************
  * NAMESPACE
  **************************************************************************************/
 
-namespace reg {
-namespace drone {
-namespace physics {
-namespace electricity {
+namespace uavcan {
+namespace primitive {
+namespace scalar {
 
 /**************************************************************************************
  * CLASS DECLARATION
  **************************************************************************************/
 
 template <CanardPortID ID>
-class Power_0_1
+class Integer32_1_0
 {
 
 public:
 
-  reg_drone_physics_electricity_Power_0_1 data;
+  uavcan_primitive_scalar_Integer32_1_0 data;
 
   static constexpr CanardPortID       PORT_ID          = ID;
-  static constexpr size_t             MAX_PAYLOAD_SIZE = reg_drone_physics_electricity_Power_0_1_SERIALIZATION_BUFFER_SIZE_BYTES_;
+  static constexpr size_t             MAX_PAYLOAD_SIZE = uavcan_primitive_scalar_Integer32_1_0_SERIALIZATION_BUFFER_SIZE_BYTES_;
   static constexpr CanardTransferKind TRANSFER_KIND    = CanardTransferKindMessage;
 
-
-  Power_0_1()
+  Integer32_1_0()
   {
-    reg_drone_physics_electricity_Power_0_1_initialize_(&data);
+    uavcan_primitive_scalar_Integer32_1_0_initialize_(&data);
   }
 
-  Power_0_1(Power_0_1 const & other)
+  Integer32_1_0(Integer32_1_0 const & other)
   {
     memcpy(&data, &other.data, sizeof(data));
   }
 
-  static Power_0_1 deserialize(CanardTransfer const & transfer)
+  static Integer32_1_0 deserialize(CanardTransfer const & transfer)
   {
-    Power_0_1 b;
+    Integer32_1_0<ID> b;
     size_t inout_buffer_size_bytes = transfer.payload_size;
-    reg_drone_physics_electricity_Power_0_1_deserialize_(&b.data, (uint8_t *)(transfer.payload), &inout_buffer_size_bytes);
+    uavcan_primitive_scalar_Integer32_1_0_deserialize_(&b.data, (uint8_t *)(transfer.payload), &inout_buffer_size_bytes);
     return b;
   }
 
   size_t serialize(uint8_t * payload) const
   {
-    size_t inout_buffer_size_bytes = Power_0_1::MAX_PAYLOAD_SIZE;
-    return (reg_drone_physics_electricity_Power_0_1_serialize_(&data, payload, &inout_buffer_size_bytes) < NUNAVUT_SUCCESS) ? 0 : inout_buffer_size_bytes;
+    size_t inout_buffer_size_bytes = Integer32_1_0::MAX_PAYLOAD_SIZE;
+    return (uavcan_primitive_scalar_Integer32_1_0_serialize_(&data, payload, &inout_buffer_size_bytes) < NUNAVUT_SUCCESS) ? 0 : inout_buffer_size_bytes;
   }
 };
 
@@ -71,9 +69,8 @@ public:
  * NAMESPACE
  **************************************************************************************/
 
-} /* reg */
-} /* drone */
-} /* physics */
-} /* electricity */
+} /* uavcan */
+} /* primitive */
+} /* scalar */
 
-#endif /* ARDUINO_UAVCAN_TYPES_REG_DRONE_PHYSICS_ELECTRICITY_POWER_0_1_H_ */
+#endif /* ARDUINO_UAVCAN_TYPES_UAVCAN_PRIMITIVE_SCALAR_INTEGER32_1_0_HPP_ */
