+++ src/wrappers/uavcan/primitive/scalar/Integer8_1_0.hpp
@@ -5,8 +5,8 @@
  * Contributors: https://github.com/107-systems/107-Arduino-UAVCAN/graphs/contributors.
  */
 
-#ifndef ARDUINO_UAVCAN_TYPES_UAVCAN_PRIMITIVE_SCALAR_REAL16_1_0_HPP_
-#define ARDUINO_UAVCAN_TYPES_UAVCAN_PRIMITIVE_SCALAR_REAL16_1_0_HPP_
+#ifndef ARDUINO_UAVCAN_TYPES_UAVCAN_PRIMITIVE_SCALAR_INTEGER8_1_0_HPP_
+#define ARDUINO_UAVCAN_TYPES_UAVCAN_PRIMITIVE_SCALAR_INTEGER8_1_0_HPP_
 
 /**************************************************************************************
  * INCLUDE
@@ -14,7 +14,7 @@
 
 #include <libcanard/canard.h>
 
-#include "Real16_1_0.h"
+#include <types/uavcan/primitive/scalar/Integer8_1_0.h>
 
 /**************************************************************************************
  * NAMESPACE
@@ -29,39 +29,39 @@ namespace scalar {
  **************************************************************************************/
 
 template <CanardPortID ID>
-class Real16_1_0
+class Integer8_1_0
 {
 
 public:
 
-  uavcan_primitive_scalar_Real16_1_0 data;
+  uavcan_primitive_scalar_Integer8_1_0 data;
 
   static constexpr CanardPortID       PORT_ID          = ID;
-  static constexpr size_t             MAX_PAYLOAD_SIZE = uavcan_primitive_scalar_Real16_1_0_SERIALIZATION_BUFFER_SIZE_BYTES_;
+  static constexpr size_t             MAX_PAYLOAD_SIZE = uavcan_primitive_scalar_Integer8_1_0_SERIALIZATION_BUFFER_SIZE_BYTES_;
   static constexpr CanardTransferKind TRANSFER_KIND    = CanardTransferKindMessage;
 
-  Real16_1_0()
+  Integer8_1_0()
   {
-    uavcan_primitive_scalar_Real16_1_0_initialize_(&data);
+    uavcan_primitive_scalar_Integer8_1_0_initialize_(&data);
   }
 
-  Real16_1_0(Real16_1_0 const & other)
+  Integer8_1_0(Integer8_1_0 const & other)
   {
     memcpy(&data, &other.data, sizeof(data));
   }
 
-  static Real16_1_0 deserialize(CanardTransfer const & transfer)
+  static Integer8_1_0 deserialize(CanardTransfer const & transfer)
   {
-    Real16_1_0<ID> b;
+    Integer8_1_0<ID> b;
     size_t inout_buffer_size_bytes = transfer.payload_size;
-    uavcan_primitive_scalar_Real16_1_0_deserialize_(&b.data, (uint8_t *)(transfer.payload), &inout_buffer_size_bytes);
+    uavcan_primitive_scalar_Integer8_1_0_deserialize_(&b.data, (uint8_t *)(transfer.payload), &inout_buffer_size_bytes);
     return b;
   }
 
   size_t serialize(uint8_t * payload) const
   {
-    size_t inout_buffer_size_bytes = Real16_1_0::MAX_PAYLOAD_SIZE;
-    return (uavcan_primitive_scalar_Real16_1_0_serialize_(&data, payload, &inout_buffer_size_bytes) < NUNAVUT_SUCCESS) ? 0 : inout_buffer_size_bytes;
+    size_t inout_buffer_size_bytes = Integer8_1_0::MAX_PAYLOAD_SIZE;
+    return (uavcan_primitive_scalar_Integer8_1_0_serialize_(&data, payload, &inout_buffer_size_bytes) < NUNAVUT_SUCCESS) ? 0 : inout_buffer_size_bytes;
   }
 };
 
@@ -73,4 +73,4 @@ public:
 } /* primitive */
 } /* scalar */
 
-#endif /* ARDUINO_UAVCAN_TYPES_UAVCAN_PRIMITIVE_SCALAR_REAL16_1_0_HPP_ */
+#endif /* ARDUINO_UAVCAN_TYPES_UAVCAN_PRIMITIVE_SCALAR_INTEGER8_1_0_HPP_ */
