+++ src/wrappers/uavcan/primitive/scalar/Real32_1_0.hpp
@@ -5,8 +5,8 @@
  * Contributors: https://github.com/107-systems/107-Arduino-UAVCAN/graphs/contributors.
  */
 
-#ifndef ARDUINO_UAVCAN_TYPES_UAVCAN_PRIMITIVE_SCALAR_BIT_1_0_HPP_
-#define ARDUINO_UAVCAN_TYPES_UAVCAN_PRIMITIVE_SCALAR_BIT_1_0_HPP_
+#ifndef ARDUINO_UAVCAN_TYPES_UAVCAN_PRIMITIVE_SCALAR_REAL32_1_0_HPP_
+#define ARDUINO_UAVCAN_TYPES_UAVCAN_PRIMITIVE_SCALAR_REAL32_1_0_HPP_
 
 /**************************************************************************************
  * INCLUDE
@@ -14,7 +14,7 @@
 
 #include <libcanard/canard.h>
 
-#include "Bit_1_0.h"
+#include <types/uavcan/primitive/scalar/Real32_1_0.h>
 
 /**************************************************************************************
  * NAMESPACE
@@ -29,39 +29,39 @@ namespace scalar {
  **************************************************************************************/
 
 template <CanardPortID ID>
-class Bit_1_0
+class Real32_1_0
 {
 
 public:
 
-  uavcan_primitive_scalar_Bit_1_0 data;
+  uavcan_primitive_scalar_Real32_1_0 data;
 
   static constexpr CanardPortID       PORT_ID          = ID;
-  static constexpr size_t             MAX_PAYLOAD_SIZE = uavcan_primitive_scalar_Bit_1_0_SERIALIZATION_BUFFER_SIZE_BYTES_;
+  static constexpr size_t             MAX_PAYLOAD_SIZE = uavcan_primitive_scalar_Real32_1_0_SERIALIZATION_BUFFER_SIZE_BYTES_;
   static constexpr CanardTransferKind TRANSFER_KIND    = CanardTransferKindMessage;
 
-  Bit_1_0()
+  Real32_1_0()
   {
-    uavcan_primitive_scalar_Bit_1_0_initialize_(&data);
+    uavcan_primitive_scalar_Real32_1_0_initialize_(&data);
   }
 
-  Bit_1_0(Bit_1_0 const & other)
+  Real32_1_0(Real32_1_0 const & other)
   {
     memcpy(&data, &other.data, sizeof(data));
   }
 
-  static Bit_1_0 deserialize(CanardTransfer const & transfer)
+  static Real32_1_0 deserialize(CanardTransfer const & transfer)
   {
-    Bit_1_0<ID> b;
+    Real32_1_0<ID> b;
     size_t inout_buffer_size_bytes = transfer.payload_size;
-    uavcan_primitive_scalar_Bit_1_0_deserialize_(&b.data, (uint8_t *)(transfer.payload), &inout_buffer_size_bytes);
+    uavcan_primitive_scalar_Real32_1_0_deserialize_(&b.data, (uint8_t *)(transfer.payload), &inout_buffer_size_bytes);
     return b;
   }
 
   size_t serialize(uint8_t * payload) const
   {
-    size_t inout_buffer_size_bytes = Bit_1_0::MAX_PAYLOAD_SIZE;
-    return (uavcan_primitive_scalar_Bit_1_0_serialize_(&data, payload, &inout_buffer_size_bytes) < NUNAVUT_SUCCESS) ? 0 : inout_buffer_size_bytes;
+    size_t inout_buffer_size_bytes = Real32_1_0::MAX_PAYLOAD_SIZE;
+    return (uavcan_primitive_scalar_Real32_1_0_serialize_(&data, payload, &inout_buffer_size_bytes) < NUNAVUT_SUCCESS) ? 0 : inout_buffer_size_bytes;
   }
 };
 
@@ -73,4 +73,4 @@ public:
 } /* primitive */
 } /* scalar */
 
-#endif /* ARDUINO_UAVCAN_TYPES_UAVCAN_PRIMITIVE_SCALAR_BIT_1_0_HPP_ */
+#endif /* ARDUINO_UAVCAN_TYPES_UAVCAN_PRIMITIVE_SCALAR_REAL32_1_0_HPP_ */
